remove @Slf4j

+++ bard-form-common/src/main/java/org/bardframework/form/field/filter/TimeFilterField.java
@@ -14,8 +14,8 @@ import java.time.temporal.ChronoUnit;
 @Setter
 @ToString
 public class TimeFilterField extends InputField<LocalTimeFilter> {
-    private Long minLength;
-    private Long maxLength;
+    private Integer minLength;
+    private Integer maxLength;
     private ChronoUnit lengthUnit;
     private LocalTime minValue;
     private LocalTime maxValue;
